refactor(account): Move `loadResources(...)` to `loader` package
